// src/home/types/nutrition-label-recognition.type.ts
export type NutritionLabelRecognition = {
  unit: number;
  weight: number;
  calories: number;
  carbs: number | null;
  sugars: number | null;
  sugar_alchol: number | null;
  dietary_fiber: number | null;
  protein: number | null;
  fat: number | null;
  sat_fat: number | null;
  trans_fat: number | null;
  un_sat_fat: number | null;
  sodium: number | null;
  caffeine: number | null;
  potassium: number | null;
  cholesterol: number | null;
  alcohol: number | null;
};